# Optimize sourcemap processing