--- a/compiler/lib/source_map.ml
+++ b/compiler/lib/source_map.ml
@@ -57,7 +57,6 @@ let gen_line = function
 let gen_col = function
   | Gen { gen_col; _ } | Gen_Ori { gen_col; _ } | Gen_Ori_Name { gen_col; _ } -> gen_col
 
-
 module Line_edits = struct
   type action =
     | Keep
@@ -75,7 +74,6 @@ module Line_edits = struct
 
   let pp fmt = Format.(pp_print_list pp_action fmt)
 end
-
 
 module Mappings = struct
   type t = Uninterpreted of string [@@unboxed]
@@ -579,7 +577,6 @@ let empty ~filename =
   ; mappings = Mappings.empty
   }
 
-
 let concat ~file ~sourceroot s1 s2 =
   if not (Int.equal s1.version s2.version)
   then invalid_arg "Source_map.concat: different versions";
@@ -695,15 +692,15 @@ let merge = function
 
 (* IO *)
 
+let rewrite_path path =
+  if Filename.is_relative path
+  then path
+  else
+    match Build_path_prefix_map.get_build_path_prefix_map () with
+    | Some map -> Build_path_prefix_map.rewrite map path
+    | None -> path
+
 let json t =
-  let rewrite_path path =
-    if Filename.is_relative path
-    then path
-    else
-      match Build_path_prefix_map.get_build_path_prefix_map () with
-      | Some map -> Build_path_prefix_map.rewrite map path
-      | None -> path
-  in
   let stringlit s = `Stringlit (Yojson.Safe.to_string (`String s)) in
   `Assoc
     [ "version", `Intlit (string_of_int t.version)
@@ -764,7 +761,7 @@ let list_stringlit_opt name rest =
     | _ -> invalid ()
   with Not_found -> None
 
-let of_json (json : Yojson.Raw.t) =
+let standard_map_of_json (json : Yojson.Raw.t) =
   match json with
   | `Assoc (("version", `Intlit version) :: rest) when int_of_string version = 3 ->
       let string name json = Option.map ~f:string_of_stringlit (stringlit name json) in
@@ -808,8 +805,116 @@ let of_json (json : Yojson.Raw.t) =
       }
   | _ -> invalid ()
 
-let of_string s = of_json (Yojson.Raw.from_string s)
-
 let to_string m = Yojson.Raw.to_string (json m)
 
 let to_file m file = Yojson.Raw.to_file file (json m)
+
+module Index = struct
+  type offset =
+    { gen_line : int
+    ; gen_column : int
+    }
+
+  (* Type synonym to avoid confusion between toplevel [t] and this submodule's [t]. *)
+  type map = t
+
+  type t =
+    { version : int
+    ; file : string
+    ; sections : (offset * [ `Map of map ]) list
+    }
+
+  let json t =
+    let stringlit s = `Stringlit (Yojson.Safe.to_string (`String s)) in
+    `Assoc
+      [ "version", `Intlit (string_of_int t.version)
+      ; "file", stringlit (rewrite_path t.file)
+      ; ( "sections"
+        , `List
+            (List.map
+               ~f:(fun ({ gen_line; gen_column }, `Map sm) ->
+                 `Assoc
+                   [ ( "offset"
+                     , `Assoc
+                         [ "line", `Intlit (string_of_int gen_line)
+                         ; "column", `Intlit (string_of_int gen_column)
+                         ] )
+                   ; "map", json sm
+                   ])
+               t.sections) )
+      ]
+
+  let intlit ~errmsg name json =
+    match List.assoc name json with
+    | `Intlit i -> int_of_string i
+    | _ -> invalid_arg errmsg
+    | exception Not_found -> invalid_arg errmsg
+
+  let section_of_json : Yojson.Raw.t -> offset * [ `Map of map ] = function
+    | `Assoc json ->
+        let offset =
+          match List.assoc "offset" json with
+          | `Assoc fields ->
+              let gen_line =
+                intlit
+                  "line"
+                  fields
+                  ~errmsg:
+                    "Source_map_io.Index.of_json: field 'line' absent or invalid from \
+                     section"
+              in
+              let gen_column =
+                intlit
+                  "column"
+                  fields
+                  ~errmsg:
+                    "Source_map_io.Index.of_json: field 'column' absent or invalid from \
+                     section"
+              in
+              { gen_line; gen_column }
+          | _ ->
+              invalid_arg "Source_map_io.Index.of_json: 'offset' field of unexpected type"
+        in
+        (match List.assoc "url" json with
+        | _ ->
+            invalid_arg
+              "Source_map_io.Index.of_json: URLs in index maps are not currently \
+               supported"
+        | exception Not_found -> ());
+        let map =
+          try standard_map_of_json (List.assoc "map" json) with
+          | Not_found -> invalid_arg "Source_map_io.Index.of_json: field 'map' absent"
+          | Invalid_argument _ ->
+              invalid_arg "Source_map_io.Index.of_json: invalid sub-map object"
+        in
+        offset, `Map map
+    | _ -> invalid_arg "Source_map_io.Index.of_json: section of unexpected type"
+
+  let of_json = function
+    | `Assoc fields -> (
+        let string name json = Option.map ~f:string_of_stringlit (stringlit name json) in
+        let file = string "file" fields in
+        match List.assoc "sections" fields with
+        | `List sections ->
+            let sections = List.map ~f:section_of_json sections in
+            { version = 3; file = Option.value file ~default:""; sections }
+        | _ -> invalid_arg "Source_map_io.Index.of_json: `sections` is not an array"
+        | exception Not_found ->
+            invalid_arg "Source_map_io.Index.of_json: no `sections` field")
+    | _ -> invalid_arg "Source_map_io.of_json: map is not an object"
+
+  let to_string m = Yojson.Raw.to_string (json m)
+
+  let to_file m file = Yojson.Raw.to_file file (json m)
+end
+
+let of_json = function
+  | `Assoc fields as json -> (
+      match List.assoc "sections" fields with
+      | _ -> `Index (Index.of_json json)
+      | exception Not_found -> `Standard (standard_map_of_json json))
+  | _ -> invalid_arg "Source_map_io.of_json: map is not an object"
+
+let of_string s = of_json (Yojson.Raw.from_string s)
+
+let of_file f = of_json (Yojson.Raw.from_file f)