--- a/compiler/lib/source_map.mli
+++ b/compiler/lib/source_map.mli
@@ -44,6 +44,19 @@ type map =
       ; ori_name : int
       }
 
+module Line_edits : sig
+  type action =
+    | Keep
+    | Drop
+    | Add of { count : int }
+
+  val pp_action : Format.formatter -> action -> unit
+
+  type t = action list
+
+  val pp : Format.formatter -> t -> unit
+end
+
 module Mappings : sig
   type t
 
@@ -54,17 +67,31 @@ module Mappings : sig
   (** By default, mappings are left uninterpreted, since many operations can be
       performed efficiently directly on the encoded form. Therefore this
       function is mostly a no-op and very cheap. It does not perform any
-      validation of its argument, unlike {!val:decode}. It is guaranteed that
-      {!val:of_string} and {!val:to_string} are inverse functions. *)
+      validation of its argument, unlike {!val:edit} or {!val:decode}. It is
+      guaranteed that {!val:of_string} and {!val:to_string} are inverse
+      functions. *)
 
   val decode : t -> map list
-  (** Parse the mappings. *)
+  (** Parse the mappings. Prefer using the more efficient {!val:edit} on the
+      uninterpreted form when applicable. *)
 
   val encode : map list -> t
 
   val to_string : t -> string
   (** Returns the mappings as a string in the Source map v3 format. This
       function is mostly a no-op and is very cheap. *)
+
+  val edit : strict:bool -> t -> Line_edits.t -> t
+  (** [edit ~strict mappings edits] applies line edits [edits] to [mappings] in
+      order and returns the result. If [strict], then the number of {!const:Keep} and
+      {!const:Drop} elements in [edits] should be the same or lesser than the
+      number of generated lines covered by [mappings]. If the number of
+      {!const:Line_edits.Keep} and {!const:Line_edits.Drop} actions is lesser
+      than the number of lines in the domain of [mappings], only the lines
+      affected by an edit are included in the result. If [strict] is false,
+      there may be more edit actions than generated lines. In which case, a
+      [Keep] that does not correspond to a line just adds a line without
+      mappings and a [Drop] without a corresponding line does nothing. *)
 end
 
 type t =
@@ -85,6 +112,12 @@ val filter_map : t -> f:(int -> int option) -> t
 val merge : t list -> t option
 
 val empty : filename:string -> t
+
+val concat : file:string -> sourceroot:string option -> t -> t -> t
+(** If [s1] encodes a mapping for a generated file [f1], and [s2] for a
+    generated file [f2], then [concat ~file ~sourceroot s1 s2] encodes the
+    union of these mappings for the concatenation of [f1] and [f2], with name
+    [file] and source root [sourceroot). *)
 
 val to_string : t -> string