@rickyvetter 2024-05-24
I just tested this change on an incremental build for one of our largest executables and am getting this timing for `js_of_ocaml link`.

Pre: 2.469s
Post: 1.301s

Very cool!