@OlivierNicole 2024-07-30
> My understanding is that the Line_edits change is what gives use the most part of the improvement. Could we achieve similar perf using sections (Index sourcemap) ?

I hoped so initially, but it turns out that `Link_js` may remove or add metadata comments in the middle of a compilation unit, which pretty much killed my hope of simply using index maps.