@OlivierNicole 2024-07-31
Never mind, apparently I had started a review at some point in the past and never completed it, and apparently Github considers all your comments henceforth as pending