@hhugo 2024-08-02
> Just to make sure we are on the same page, my intention is to wait for #1640 to be merged, rebase on it, and then address the rest of your new review (thanks!).

What do you think about the following, on top of #1640.
- drop the indexed sourcemap for now (it seems it's not used)
- first merge the delayed/on demand decoding that replace the sourcemap change from #1654  
- keep this PR for "edit un-decoded mapping" only.