@hhugo 2024-09-05
I'm not very happy with the callback based approach used on `Link_js`. Without thinking much about it, I would assume that one could reconstruct the edit instruction inside `let copy ic oc = ...` only.

Maybe we could move the handling/parsing of Sourcemap index into a separate PR to remove some noise in here.