@hhugo 2024-09-05
Regarding the representations of edits. We currently have per line instructions. Have you considered using a different approach (closer to what's done on master with `reloc`) using "copy n lines starting at pos n into position m" ?