@OlivierNicole 2024-10-10
> Regarding the representations of edits. We currently have per line instructions. Have you considered using a different approach (closer to what's done on master with `reloc`) using "copy n lines starting at pos n into position m" ?

Actually, what I had considered was to have the edits represented as a list of “copy / delete n lines” rather than a long list of “copy / delete 1 line”.

What you suggest would make `Source_map.Mappings.edit` no longer linear. In your suggestion, some sorting is involved, whereas in what is currently done, the destination line is implicitly represented by the position of the edit in the list of edits.