@hhugo 2024-10-15
> > My understanding is that the Line_edits change is what gives use the most part of the improvement. Could we achieve similar perf using sections (Index sourcemap) ?
> 
> I hoped so initially, but it turns out that `Link_js` may remove or add metadata comments in the middle of a compilation unit, which pretty much killed my hope of simply using index maps.

Can we maybe revisit this approach. Maybe we could  teach link_js to copy blocks of lines that do not mess up with sourcemap ? If we really want to drop comments, we could emit an empty line instead. We would probably need to emit Index map when compiling cma files so that we have one sourcemap per unit and we can easily drop dead units when linking.