@hhugo 2024-10-17
I've opened #1715 instead. It removes the need to change encoded mappings.